融合过滤 业务模型关系属性id类型变更

+++ kgms-sdk/src/main/java/ai/plantdata/kgcloud/sdk/req/app/algorithm/BusinessRelationBean.java
@@ -19,7 +19,7 @@ public class BusinessRelationBean {
     private String id;
     private Long from;
     private Long to;
-    private Integer attId;
+    private Long attId;
     private String attName;
     private List<String> startTime;
     private List<String> endTime;
@@ -71,7 +71,7 @@ public class BusinessRelationBean {
     }
 
 
-    public BusinessRelationBean(String id, Long from, Long to, Integer attId) {
+    public BusinessRelationBean(String id, Long from, Long to, Long attId) {
         this.id = id;
         this.from = from;
         this.to = to;
@@ -118,11 +118,11 @@ public class BusinessRelationBean {
         this.to = to;
     }
 
-    public Integer getAttId() {
+    public Long getAttId() {
         return attId;
     }
 
-    public void setAttId(Integer attId) {
+    public void setAttId(Long attId) {
         this.attId = attId;
     }
